Change watch() semantics to not trigger on revisions individually
The main change is that we don't start the new transaction at the
revision discovered via the watch notification. This means that after
receiving a notification, we always ask the database again for what
the "current" revision is, reducing unnecessary churn if there are
many updates to a key.

This also means that there is no reason to not empty all queues
completely every time watch() gets called. This would happen anyway on
the next call, but this wastes slightly less memory.