Working on list watches
One of the features that need a bit of busy work because of the
possible interactions. We *can* watch a range, but that will also tell
us about key updates in range. This is actually great for our
purposes, because we will often list *and* read keys, yet this means
that we

 1. need to carefully consider exactly what watchers we need to
    instantiate in every situation, as "list" watchers will often
    make "get" watchers redundant

 2. pay attention to the watch events coming back and figure out
    whether the ones not matching straight-up "get" requests would
    actually change the result of "list_keys"

It is not terribly complicated, but requires a bit of house keeping.