Support for single-key delete in transactions
This is fairly straightforward to support. Recursive or prefix deletes
are a bit tricker because we need to keep track of anything that might
enter the range after the fact. Clearly not impossible - we are already
doing that for list_keys, and in fact a simple way of doing this would
simply be to use list_key() and iterate delete() over it. Not really
very efficient, but fairly obviously correct...