Continued work on configuration database [ORCA-89] (#5)
* Small schema updates

Clarify how receive addresses and commands get communicated through
the configuration database

* Change watch() semantics to not trigger on revisions individually

The main change is that we don't start the new transaction at the
revision discovered via the watch notification. This means that after
receiving a notification, we always ask the database again for what
the "current" revision is, reducing unnecessary churn if there are
many updates to a key.

This also means that there is no reason to not empty all queues
completely every time watch() gets called. This would happen anyway on
the next call, but this wastes slightly less memory.

* Check for trailing slash in a couple more cases

* Correctly implement retry counter

Must reset the counter on "successful" transaction loops

* Working on list watches

One of the features that need a bit of busy work because of the
possible interactions. We *can* watch a range, but that will also tell
us about key updates in range. This is actually great for our
purposes, because we will often list *and* read keys, yet this means
that we

 1. need to carefully consider exactly what watchers we need to
    instantiate in every situation, as "list" watchers will often
    make "get" watchers redundant

 2. pay attention to the watch events coming back and figure out
    whether the ones not matching straight-up "get" requests would
    actually change the result of "list_keys"

It is not terribly complicated, but requires a bit of house keeping.

* Improve unit test coverage

Actually test watching a list of keys. We need to introduce a way to
have blocking queries in test cases for this.

* Support for single-key delete in transactions

This is fairly straightforward to support. Recursive or prefix deletes
are a bit tricker because we need to keep track of anything that might
enter the range after the fact. Clearly not impossible - we are already
doing that for list_keys, and in fact a simple way of doing this would
simply be to use list_key() and iterate delete() over it. Not really
very efficient, but fairly obviously correct...

* High-level API using transactions

This works much better I think. Entities are purely for holding
and validating dictionaries (JSON objects), and implement no
database tasks of their own. The idea is to have all of that in the
Transaction class so that the database traffic is entirely obvious.

* More transaction test coverage

Properly test delete and lease functions for the transaction
interface. This uncovered two couple small bugs.

* Backend documentation

Especialy add documentation to transaction methods. Also add a lengthy
TODO comment about things we might want to do in the future.

* Processing block ownership

The last really critical feature we need - this implements a lock /
leader election as well as a basic liveness check. Idea is that we
guarantee that only one processing block controller "owns" a
processing block at any given time.

* Add command line interface

Including appropriate files to make this pip-installable

* Add docopt dependency

* Add example workflow

This workflow simply claims a processing block from the configuration
and holds it until the processing block gets removed or ownerhip gets
revoked externally.

* Fix linter problems

Actually managed to spot a couple genuine oversights.

* Add config_db tests to CI script

* Rename CI test job

* Attempt to make etcd3 components available for test

* Try latest etcd3, and installing "wheel"

* Install wheel as a separate step?

* Use pipenv

* Attempt to find etcd at different host

* Set alias name explicitly

* Add simple CLI test cases

* Add config_db documentation

* Bind etcd3 to 0.0.0.0, use correct host for CLI test