Add functionality to SDPSubarray device receiveAddresses attribute [ORCA-106] (#45)
Add receiveAddresses functionality to SDPSubarray.

This reads a channel - FSP link map from a CSP subarray and generates a JSON mapping for the receiveAddresses according to the descriptions found at
<https://confluence.skatelescope.org/pages/viewpage.action?pageId=74716479>
and
<https://confluence.skatelescope.org/display/SE/Mid.CBF+Output+Links+and+Destination+Addresses+JSON+Schema>

Added a feature toggle attribute to read or ignore then cbfOutLink
CSP Subarray attribute. If set to False this populates receiveAddresses
with a basic hardcoded test response.