Make all SDP pods handle SIGTERM
This is to make them terminate in a graceful fashion when Kubernetes
is trying to shut SDP pods down. For now this just executes exit(),
which is is still strictly better than waiting for SIGKILL.

The obvious part here is to install a signal handler. Less obvious is
that the Dockerfiles need to be adjusted to use the exec instead of
the shell notation (CMD x y z -> CMD ["x", "y", "z"]) as otherwise
SIGTERM is only delivered to the enclosing shell process instead of
the Python program.