Changing how python scripts are launched
On the one hand, instead of hardcoding a name for the python interpreter
we use sys.executable, which is our interpreter. And on the second,
instead of starting the python programs as scripts we start them as
modules, as this will allow them to be better packaged and modularized
in the future.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>