Fix Tango device version number import issue
Fixes an issue where the import of the server version number from
release.py fails either when running unit tests or starting the
Tango device server. This seems to be a combination of python3
import rules and the way the Tango device server name is resolved
which limits the way the device server can be started. This fix
adds the SDPMaster / SDPSubarray source folder to the import
search path (sys.path). This is hardly an ideal 'fix' but should be a
reasonable work around until the issue can be addressed properly!